Add uint16 output
